superbeast

+++ lib/roro/cli/templates/greenfield/greenfield-entrypoint.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
 # Prefix `bundle` with `exec` so unicorn shuts down gracefully on SIGTERM (i.e. `docker stop`)
-exec mv -f config/database.yml.example config/database.yml
+
+file="config/database.yml.example"
+if [ ! -f "$file" ]
+then
+    exec mv -f config/database.yml.example config/database.yml
+fi
 exec bundle exec rails s -p 3000 -b '0.0.0.0'
